Handle Installation to Virtual Environments for macOS, and Windows (#34)
Added a bdist_wheel command inside of a try block.  This ensure that
normal system wide installations, which don't include the wheel module,
are not impacted.

For virtual environment the wheel module will be present, and this will
override it for the virtual environment installation.  This should
ensure that jars are downloaded for both standard, and virtual
environment installations.